requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
sqlalchemy==2.0.23
opencv-python==4.8.1.78
moviepy==1.0.3
python-dotenv==1.0.0
pyjwt==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
werkzeug==2.3.7
numpy==1.26.1